Specify a platform (linux/amd64) for Debian base image.
The version of Operator used in the image is compiled for x86-64.
However, the Debian image is available for multiple platforms and there
is ambiguity unless --platform is specified.

This came up now because I tried to build this image on Apple silicon
for the first time. Docker defaults to the host platform's architecture
and was pulling a Debian image which can't run the server.

I'll probably build Operator for this Apple silicon eventually, but
even then it probably makes sense to default to x86-64 for this image.